+++ src/client/js/startup.js
@@ -3,10 +3,10 @@
  */
 
 Meteor.startup(() => {
-    if( Meteor.cookieManager && pwixI18n._conf.storePreferredLanguage ){
-        Meteor.cookieManager.publish({
+    if( Meteor.CookieManager ){
+        Meteor.CookieManager.publish({
+            responsible: COOKIE_RESPONSIBLE,
             name: COOKIE_PREFERRED_LANGUAGE,
-            responsible: 'pwix:i18n',
             description: pwixI18n.label( I18N, 'cookies.preferred_language' ),
             category: 'CM_CAT_FUNCTIONALS',
             lifetime: pwixI18n.label( I18N, 'cookies.illimited' ),