+++ src/common/js/constants.js
@@ -33,4 +33,6 @@ NOLANG = 'NOLANG';
 I18N = 'pwixI18n:i18n:namespace';
 
 // the cookie name for the preferred language, defined here to be available anywhere
-COOKIE_PREFERRED_LANGUAGE = 'pwix:i18n/preferred_language';
+COOKIE_RESPONSIBLE = 'pwix:i18n';
+COOKIE_PREFERRED_LANGUAGE = 'preferred_language';
+COOKIE_PREFERRED_LANGUAGE_STORE = 'pwix:i18n/preferred_language';