+++ src/common/js/store.js
@@ -7,26 +7,34 @@
  * If application doesn't use a cookie manager, then we consider that cookies are allowed.
  */
 
-pwixI18n._storeGet = function( key ){
-    let result =  null;
-    if( Meteor.isClient && pwixI18n._conf.storePreferredLanguage ){
-        let enabled = true;
-        if( Meteor.cookieManager ){
-            enabled = Meteor.cookieManager.isEnabled( key );
+pwixI18n._store = {
+    /**
+     * @summary read from localStore
+     */
+    get( key ){
+        let result =  null;
+        if( Meteor.isClient && pwixI18n._conf.storePreferredLanguage ){
+            let enabled = true;
+            if( Meteor.CookieManager ){
+                enabled = Meteor.CookieManager.isEnabled( key );
+            }
+            result = enabled ? localStorage.getItem( key ) : null;
         }
-        result = enabled ? localStorage.getItem( key ) : null;
-    }
-    return result;
-};
-
-pwixI18n._storeSet = function( key, value ){
-    if( Meteor.isClient && pwixI18n._conf.storePreferredLanguage ){
-        let enabled = true;
-        if( Meteor.cookieManager ){
-            enabled = Meteor.cookieManager.isEnabled( key );
-        }
-        if( enabled ){
-            localStorage.setItem( key, value );
+        return result;
+    },
+    
+    /**
+     * @summary write in localStore
+     */
+    set( key, value ){
+        if( Meteor.isClient && pwixI18n._conf.storePreferredLanguage ){
+            let enabled = true;
+            if( Meteor.CookieManager ){
+                enabled = Meteor.CookieManager.isEnabled( key );
+            }
+            if( enabled ){
+                localStorage.setItem( key, value );
+            }
         }
     }
 };